CMake: Fix finding ncurses headers + new implicit linker search dirs feature
NCurses:

The ncurses include path was being miscalculated because, in find_path, NAMES
must not include folders as the folder prefix is not stored in VAR.

Implicit Linker Search Dirs:

We ask the linker directly for its search dirs rather than going through
the compiler.  This is necessary when the search dirs for the compiler and
those for the linker differ (e.g. when using conda gcc and the system ld).

This difference matters because CMake removes 'implicit directories' from
commandlines, and if it removes ones that the linker needs to be told about
then the linker will not be able to find shared objects.

Two cases that need this are -Wl,-rpath,<paths> and -Wl,-rpath-link,<paths>
but also any time that the compiler passes paths directly to the linker.